+++ tests/framework/data/ActiveDataProviderTest.php
@@ -35,6 +35,7 @@ abstract class ActiveDataProviderTest extends DatabaseTestCase
     public function testActiveQuery()
     {
         $provider = new ActiveDataProvider([
+            'db' => $this->getConnection(),
             'query' => Order::find()->orderBy('id'),
         ]);
         $orders = $provider->getModels();
@@ -45,6 +46,7 @@ abstract class ActiveDataProviderTest extends DatabaseTestCase
         $this->assertEquals([1, 2, 3], $provider->getKeys());
 
         $provider = new ActiveDataProvider([
+            'db' => $this->getConnection(),
             'query' => Order::find(),
             'pagination' => [
                 'pageSize' => 2,
@@ -59,6 +61,7 @@ abstract class ActiveDataProviderTest extends DatabaseTestCase
         /* @var $customer Customer */
         $customer = Customer::findOne(2);
         $provider = new ActiveDataProvider([
+            'db' => $this->getConnection(),
             'query' => $customer->getOrders(),
         ]);
         $orders = $provider->getModels();
@@ -68,6 +71,7 @@ abstract class ActiveDataProviderTest extends DatabaseTestCase
         $this->assertEquals([2, 3], $provider->getKeys());
 
         $provider = new ActiveDataProvider([
+            'db' => $this->getConnection(),
             'query' => $customer->getOrders(),
             'pagination' => [
                 'pageSize' => 1,
@@ -82,6 +86,7 @@ abstract class ActiveDataProviderTest extends DatabaseTestCase
         /* @var $order Order */
         $order = Order::findOne(2);
         $provider = new ActiveDataProvider([
+            'db' => $this->getConnection(),
             'query' => $order->getItems(),
         ]);
         $items = $provider->getModels();
@@ -92,6 +97,7 @@ abstract class ActiveDataProviderTest extends DatabaseTestCase
         $this->assertEquals([3, 4, 5], $provider->getKeys());
 
         $provider = new ActiveDataProvider([
+            'db' => $this->getConnection(),
             'query' => $order->getItems(),
             'pagination' => [
                 'pageSize' => 2,
@@ -106,6 +112,7 @@ abstract class ActiveDataProviderTest extends DatabaseTestCase
         /* @var $order Order */
         $order = Order::findOne(1);
         $provider = new ActiveDataProvider([
+            'db' => $this->getConnection(),
             'query' => $order->getBooks(),
         ]);
         $items = $provider->getModels();
@@ -197,5 +204,21 @@ abstract class ActiveDataProviderTest extends DatabaseTestCase
         }
 
         $this->assertEquals(0, $pagination->getPageCount());
+    }
+
+    public function testPaginationWithUnionQuery()
+    {
+        $q1 = Item::find()->where(['category_id' => 2]);
+        $q2 = Item::find()->where(['id' => 1]);
+        $provider = new ActiveDataProvider([
+            'db' => $this->getConnection(),
+            'query' => $q1->union($q2),
+        ]);
+        $pagination = $provider->getPagination();
+        $pagination->pageSize = 2;
+        $provider->prepare();
+        $this->assertEquals(2, $pagination->getPageCount());
+        $this->assertEquals(4, $provider->getTotalCount());
+        $this->assertCount(2, $provider->getModels());
     }
 }